indigo_server: fix type cast

+++ indigo_server/indigo_server.c
@@ -1151,7 +1151,7 @@ static indigo_result change_property(indigo_device *device, indigo_client *clien
 	} else if (indigo_property_match(SERVER_WIFI_CHANNEL_PROPERTY, property)) {
 		// -------------------------------------------------------------------------------- SERVER_WIFI_CHANNEL
 		indigo_property_copy_values(SERVER_WIFI_CHANNEL_PROPERTY, property, false);
-		return execute_command(device, SERVER_WIFI_CHANNEL_PROPERTY, "s_rpi_ctrl.sh --set-wifi-channel %d", SERVER_WIFI_CHANNEL_ITEM->number.value);
+		return execute_command(device, SERVER_WIFI_CHANNEL_PROPERTY, "s_rpi_ctrl.sh --set-wifi-channel %d", (int)SERVER_WIFI_CHANNEL_ITEM->number.value);
 	} else if (indigo_property_match(SERVER_INTERNET_SHARING_PROPERTY, property)) {
 		// -------------------------------------------------------------------------------- INTERNET_SHARING
 		indigo_property_copy_values(SERVER_INTERNET_SHARING_PROPERTY, property, false);
